move custom messages to proto.h
